fix: refactor duplicate slack test mock code
